requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
